Issue #95 Initial attempt to refactor purchase delay
Wait period comparison to compare previous systime with a constant offset
and check whether it is greater than current systime.
The previous implementation did not provide consistent wait period, as the
initial purchase could be at the end of the shifted time (e.g. 7 >> 3 ==
0, but 1 tick later would be 1 (and allow a purchase). The user would then
need to wait 8 ticks for the next purchase.
The previous implementation was a very inprecise granularity, as the wait
period could only be a multiple of 2.
-e 
Original/build commit ID: 2974f57870b693d26a73e93be8412ae627577cac

+++ game_state.h
@@ -90,7 +90,7 @@ typedef struct {
     // This is used to hold the systime of the last movement.
     // In building menu, this is used to hold the systime of the
     // last item purchase, which stops accidental double purchsing
-    UINT16 last_movement_time;
+    unsigned int last_movement_time;
 
     BOOLEAN hobo_given_money;
     BOOLEAN hobo_given_beer;
